backend/requirements.txt
fastapi
uvicorn
tensorflow-cpu
xgboost
scikit-learn
shap
lime
joblib
numpy
pandas
python-multipart